feat: restructure levelset reinitialization
