+++ src/euler_fluid/reinitialize_levelset/shaders/jump_flooding/calculate_sdf.wgsl
